Add site-wide sticky apply banner
Promote the applications banner to a sticky announcement bar pinned above
the navbar on every page (previously only inside the Join Us hero).

- New ApplyBar: slim fixed top bar, renders only while APPLICATION_IS_LIVE
- Layout sets --apply-banner-h so the navbar and all page content offset by
  the bar's height (0 when the bar is hidden)
- Navbar top anchored to --apply-banner-h
- Remove the now-duplicate inline ApplyNowBanner from JoinUsHero and delete
  the unused component
